server/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
supabase
pydantic
python-multipart
reportlab
torch
transformers
Pillow
